example/requirements.txt
fastapi==0.111.0
python-dotenv==1.0.1
cryptography==42.0.8
tortoise-orm==0.21.7
asyncpg==0.29.0
aerich==0.7.2
solo-django-admin==0.1.5
psycopg2==2.9.10
